refactor: change of file for GetBoardOverview from its own file into /database/boards
