+++ server/src/database/boards.go
@@ -192,3 +192,27 @@ func (d *Database) GetBoards(userID uuid.UUID) ([]Board, error) {
 	}
 	return boards, err
 }
+
+func (d *Database) GetBoardOverview(id uuid.UUID) (Board, []BoardSession, []Column, error) {
+	var board Board
+	var sessions []BoardSession
+	var columns []Column
+	var err error
+
+	board, err = d.GetBoard(id)
+	if err != nil {
+		return Board{}, nil, nil, err
+	}
+	sessions, err = d.GetBoardSessions(id)
+	if err != nil {
+		return Board{}, nil, nil, err
+	}
+
+	columns, err = d.GetColumns(id)
+	if err != nil {
+		return Board{}, nil, nil, err
+	}
+
+	return board, sessions, columns, err
+
+}
